# backendnest/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestão Financeira Pessoal
  description: |
    API REST para autenticação e gestão de contas, categorias, transações,
    transferências, dívidas, metas e alertas.

    ## Convenção de idioma
    Os nomes dos campos seguem **camelCase em português**, pois as entidades
    do banco de dados estão em português (ex. `nome`, `senha`, `saldoInicial`).
    Alguns campos de domínio financeiro estão em espanhol (ex. `montoTotal`,
    `fechaVencimiento`, `diasAnticipacion`).

    ## Autenticação
    Todos os endpoints (exceto `/auth/register` e `/auth/login`) requerem um
    token JWT no header `Authorization: Bearer <token>`.
    O token expira em **1 hora**.

    ## Códigos de status
    - `200` — Operação bem-sucedida (GET, PATCH, DELETE)
    - `201` — Recurso criado (POST)
    - `400` — Validação de campos falhou
    - `401` — Token ausente, expirado ou inválido
    - `404` — Recurso não encontrado ou não pertence ao usuário
    - `409` — Conflito (ex. e-mail duplicado)
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido em /auth/login. Expira em 1 hora.
  schemas:
...
    CreateTransferenciaRequest:
      type: object
      required: [contaOrigemId, contaDestinoId, valor, data]
      properties:
        contaOrigemId:
          type: string
          format: uuid
          description: ID da conta de origem. Não pode ser igual a contaDestinoId.
        contaDestinoId:
          type: string
          format: uuid
          description: ID da conta de destino. Não pode ser igual a contaOrigemId.
        valor:
          type: number
          format: double
          description: Valor da transferência (decimal 12,2).
          example: 200.00
        data:
          type: string
          format: date
          description: Data da transferência.
        descricao:
          type: string
          nullable: true
          description: Descrição opcional.
        comissao:
          type: number
          format: double
          default: 0
          description: Comissão ou taxa da transferência (padrão 0).
    UpdateTransferenciaRequest:
      type: object
      description: Apenas os campos enviados serão atualizados. Não é permitido alterar as contas de origem/destino.
      properties:
        valor:
          type: number
          format: double
          description: Novo valor.
        data:
          type: string
          format: date
          description: Nova data.
        descricao:
          type: string
          description: Nova descrição.
        comissao:
          type: number
          format: double
          description: Nova comissão.
    Transferencia:
      type: object
      description: Entidade completa de uma transferência entre contas.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único.
        usuarioId:
          type: string
          format: uuid
          description: ID do usuário dono.
        contaOrigemId:
          type: string
          format: uuid
          description: ID da conta de origem.
        contaDestinoId:
          type: string
          format: uuid
          description: ID da conta de destino.
        valor:
          type: number
          format: double
          description: Valor transferido.
        data:
          type: string
          format: date
          description: Data da transferência.
        descricao:
          type: string
          nullable: true
          description: Descrição opcional.
        comissao:
          type: number
          format: double
          description: Comissão cobrada.
        moeda:
          type: string
          description: Moeda da transferência.
          example: BRL
        createdAt:
          type: string
          format: date-time
          description: Data de criação.
        updatedAt:
          type: string
          format: date-time
          description: Data da última atualização.

    # ══════════════════════════════════════
    #  DIVIDAS
    # ══════════════════════════════════════
...
    Error400:
      type: object
      description: Erro de validação. O campo `message` pode ser string ou array de strings com os erros específicos.
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Detalhe do erro de validação.
        error:
          type: string
          example: Bad Request
    Error401:
      type: object
      description: Erro de autenticação (token ausente, expirado ou inválido).
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
    Error404:
      type: object
      description: Recurso não encontrado ou não pertence ao usuário autenticado.
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Conta não encontrada
...
  responses:
...
    Unauthorized:
      description: Token JWT ausente, expirado ou inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error401'
    NotFound:
      description: Recurso não encontrado ou não pertence ao usuário autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error404'
...
paths:
...
  /transferencias:
    post:
      tags: [Transferencias]
      summary: Criar transferência
      description: |
        Cria uma transferência entre duas contas do usuário.
        Valida que ambas as contas existam e que **origem ≠ destino** (retorna 400 se forem iguais).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferenciaRequest'
      responses:
        '201':
          description: Transferência criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transferencia'
        '400':
          description: Conta de origem e destino são a mesma.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conta de origem ou destino não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
    get:
      tags: [Transferencias]
      summary: Listar transferências
      description: |
        Retorna todas as transferências do usuário autenticado,
        ordenadas por data **descendente**.
        Atualmente não suporta filtros por query params.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de transferências ordenadas por data DESC.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transferencia'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /transferencias/{id}:
    get:
      tags: [Transferencias]
      summary: Obter transferência por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da transferência.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhe da transferência.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transferencia'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Transferencias]
      summary: Atualizar transferência
      description: Atualiza os campos enviados. Não é permitido alterar as contas. Retorna a entidade atualizada.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da transferência.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransferenciaRequest'
      responses:
        '200':
          description: Transferência atualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transferencia'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Transferencias]
      summary: Excluir transferência
      description: Exclui a transferência permanentemente (hard delete). Não retorna body.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da transferência.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transferência excluída. Sem body de resposta.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════
  #  DIVIDAS
  # ════════════════════════════════════════
...